+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Add interface `ILexer`.
   - Add `Lexer.clone`.
   - `Lexer.lex` add optional parameter `expect` to limit the output token type.
+  - Add `Lexer.trimStart`.
 - Parser
   - Use `DefinitionContextBuilder` to define parser actions.
   - If ASTNode has no error, the error field will be `undefined`.